+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.0
+- added flow and prettier
+- added class viewer
+- reorganized components
+- etc
+
 # 0.1.0
 - created react app
 - added nav and basic css